:art: split async and sync method
