+++ src/EasyCaching.CSRedis/DefaultCSRedisCachingProvider.cs
@@ -1,15 +1,12 @@
 ﻿namespace EasyCaching.CSRedis
 {
-    using EasyCaching.Core;
-    using EasyCaching.Core.Internal;
-    using EasyCaching.Core.Serialization;
-    using global::CSRedis;
-    using Microsoft.Extensions.Logging;
-    using Microsoft.Extensions.Options;
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using System.Threading.Tasks;
+    using EasyCaching.Core;
+    using EasyCaching.Core.Serialization;
+    using global::CSRedis;
+    using Microsoft.Extensions.Logging;
 
     public partial class DefaultCSRedisCachingProvider : EasyCachingAbstractProvider
     {
@@ -105,19 +102,7 @@
 
             return _cache.Exists(cacheKey);
         }
-
-        /// <summary>
-        /// Existses the async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        public override async Task<bool> BaseExistsAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            return await _cache.ExistsAsync(cacheKey);
-        }
-
+     
         /// <summary>
         /// Flush this instance.
         /// </summary>
@@ -127,18 +112,6 @@
                 _logger?.LogInformation("Redis -- Flush");
 
             _cache.NodesServerManager.FlushDb();
-        }
-
-        /// <summary>
-        /// Flushs the async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        public override async Task BaseFlushAsync()
-        {
-            if (_options.EnableLogging)
-                _logger?.LogInformation("Redis -- FlushAsync");
-
-            await _cache.NodesServerManager.FlushDbAsync();
         }
 
         /// <summary>
@@ -251,152 +224,7 @@
 
             return result;
         }
-
-        /// <summary>
-        /// Gets all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="cacheKeys">Cache keys.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetAllAsync<T>(IEnumerable<string> cacheKeys)
-        {
-            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
-
-            var result = new Dictionary<string, CacheValue<T>>();
-
-            //maybe we should use mget here based on redis mode
-            //multiple keys may trigger `don't hash to the same slot`
-
-            foreach (var item in cacheKeys)
-            {
-                var cachedValue = await _cache.GetAsync<byte[]>(item);
-                if (cachedValue != null)
-                    result.Add(item, new CacheValue<T>(_serializer.Deserialize<T>(cachedValue), true));
-                else
-                    result.Add(item, CacheValue<T>.NoValue);
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Gets the async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="dataRetriever">Data retriever.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey, Func<Task<T>> dataRetriever, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            var result = await _cache.GetAsync<byte[]>(cacheKey);
-            if (result != null)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                var value = _serializer.Deserialize<T>(result);
-                return new CacheValue<T>(value, true);
-            }
-
-            CacheStats.OnMiss();
-
-            if (_options.EnableLogging)
-                _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-            var flag = await _cache.SetAsync($"{cacheKey}_Lock", 1, TimeSpan.FromMilliseconds(_options.LockMs).Seconds, RedisExistence.Nx);
-
-            if (!flag)
-            {
-                await Task.Delay(_options.SleepMs);
-                return await GetAsync(cacheKey, dataRetriever, expiration);
-            }
-
-            var item = await dataRetriever();
-            if (item != null)
-            {
-                await SetAsync(cacheKey, item, expiration);
-                //remove mutex key
-                await _cache.DelAsync($"{cacheKey}_Lock");
-                return new CacheValue<T>(item, true);
-            }
-            else
-            {
-                //remove mutex key
-                await _cache.DelAsync($"{cacheKey}_Lock");
-                return CacheValue<T>.NoValue;
-            }
-        }
-
-        /// <summary>
-        /// Gets the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="type">Object Type.</param>
-        public override async Task<object> BaseGetAsync(string cacheKey, Type type)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var result = await _cache.GetAsync<byte[]>(cacheKey);
-            if (result != null)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                var value = _serializer.Deserialize(result, type);
-                return value;
-            }
-            else
-            {
-                CacheStats.OnMiss();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Gets the async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var result = await _cache.GetAsync<byte[]>(cacheKey);
-            if (result != null)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                var value = _serializer.Deserialize<T>(result);
-                return new CacheValue<T>(value, true);
-            }
-            else
-            {
-                CacheStats.OnMiss();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-                return CacheValue<T>.NoValue;
-            }
-        }
-
+ 
         /// <summary>
         /// Handles the prefix of CacheKey.
         /// </summary>
@@ -469,34 +297,6 @@
         }
 
         /// <summary>
-        /// Gets the by prefix async.
-        /// </summary>
-        /// <returns>The by prefix async.</returns>
-        /// <param name="prefix">Prefix.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetByPrefixAsync<T>(string prefix)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
-
-            prefix = this.HandlePrefix(prefix);
-
-            var redisKeys = this.SearchRedisKeys(prefix);
-
-            var result = new Dictionary<string, CacheValue<T>>();
-
-            foreach (var item in redisKeys)
-            {
-                var cachedValue = await _cache.GetAsync<byte[]>(item);
-                if (cachedValue != null)
-                    result.Add(item, new CacheValue<T>(_serializer.Deserialize<T>(cachedValue), true));
-                else
-                    result.Add(item, CacheValue<T>.NoValue);
-            }
-
-            return result;
-        }
-
-        /// <summary>
         /// Gets the count.
         /// </summary>
         /// <returns>The count.</returns>
@@ -546,36 +346,9 @@
         }
 
         /// <summary>
-        /// Removes all async.
+        /// Remove cached value by prefix
         /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="cacheKeys">Cache keys.</param>
-        public override async Task BaseRemoveAllAsync(IEnumerable<string> cacheKeys)
-        {
-            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
-
-            var tasks = new List<Task<long>>();
-
-            foreach (var item in cacheKeys)
-            {
-                tasks.Add(_cache.DelAsync(item));
-            }
-
-            await Task.WhenAll(tasks);
-        }
-
-        /// <summary>
-        /// Removes the async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        public override async Task BaseRemoveAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            await _cache.DelAsync(cacheKey);
-        }
-
+        /// <param name="prefix">The prefix of cache key</param>
         public override void BaseRemoveByPrefix(string prefix)
         {
             ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
@@ -591,32 +364,6 @@
             {
                 _cache.Del(item);
             }
-        }
-
-        /// <summary>
-        /// Removes the by prefix async.
-        /// </summary>
-        /// <returns>The by prefix async.</returns>
-        /// <param name="prefix">Prefix.</param>
-        public override async Task BaseRemoveByPrefixAsync(string prefix)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
-
-            prefix = this.HandlePrefix(prefix);
-
-            if (_options.EnableLogging)
-                _logger?.LogInformation($"RemoveByPrefixAsync : prefix = {prefix}");
-
-            var redisKeys = this.SearchRedisKeys(prefix);
-
-            var tasks = new List<Task<long>>();
-
-            foreach (var item in redisKeys)
-            {
-                tasks.Add(_cache.DelAsync(item));
-            }
-
-            await Task.WhenAll(tasks);
         }
 
         /// <summary>
@@ -648,10 +395,10 @@
         /// <summary>
         /// Sets all.
         /// </summary>
-        /// <param name="value">Value.</param>
+        /// <param name="values">Value.</param>
         /// <param name="expiration">Expiration.</param>
         /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override void BaseSetAll<T>(IDictionary<string, T> value, TimeSpan expiration)
+        public override void BaseSetAll<T>(IDictionary<string, T> values, TimeSpan expiration)
         {
             //whether to use pipe based on redis mode 
             if (MaxRdSecond > 0)
@@ -660,65 +407,10 @@
                 expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
             }
 
-            foreach (var item in value)
+            foreach (var item in values)
             {
                 _cache.Set(item.Key, _serializer.Serialize(item.Value), (int)expiration.TotalSeconds);
             }
-        }
-
-        /// <summary>
-        /// Sets all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="value">Value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task BaseSetAllAsync<T>(IDictionary<string, T> value, TimeSpan expiration)
-        {
-            //whether to use pipe based on redis mode 
-            var tasks = new List<Task<bool>>();
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            foreach (var item in value)
-            {
-                tasks.Add(_cache.SetAsync(item.Key, _serializer.Serialize(item.Value), (int)expiration.TotalSeconds));
-            }
-
-            await Task.WhenAll(tasks);
-        }
-
-        /// <summary>
-        /// Sets the async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="cacheValue">Cache value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task BaseSetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            var val = _serializer.Serialize(cacheValue);
-
-            await _cache.SetAsync(
-                cacheKey,
-                val,
-                (int)expiration.TotalSeconds
-                );
         }
 
         /// <summary>
@@ -750,33 +442,10 @@
         }
 
         /// <summary>
-        /// Tries the set async.
+        /// Get the expiration of cache key
         /// </summary>
-        /// <returns>The set async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="cacheValue">Cache value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<bool> BaseTrySetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            return await _cache.SetAsync(
-                cacheKey,
-                _serializer.Serialize(cacheValue),
-                (int)expiration.TotalSeconds,
-                RedisExistence.Nx
-                );
-        }
-
+        /// <param name="cacheKey">cache key</param>
+        /// <returns>expiration</returns>
         public override TimeSpan BaseGetExpiration(string cacheKey)
         {
             ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
@@ -785,14 +454,10 @@
             return TimeSpan.FromSeconds(second);
         }
 
-        public override async Task<TimeSpan> BaseGetExpirationAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var second = await _cache.TtlAsync(cacheKey);
-            return TimeSpan.FromSeconds(second);
-        }
-
+        /// <summary>
+        /// Get te information of this provider.
+        /// </summary>
+        /// <returns></returns>
         public override ProviderInfo BaseGetProviderInfo()
         {
             return _info;
