+++ src/EasyCaching.InMemory/DefaultInMemoryCachingProvider.Async.cs
@@ -1,0 +1,306 @@
+﻿namespace EasyCaching.InMemory
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
+    using EasyCaching.Core;
+    using Microsoft.Extensions.Logging;
+
+    /// <summary>
+    /// MemoryCaching provider.
+    /// </summary>
+    public partial class DefaultInMemoryCachingProvider : EasyCachingAbstractProvider
+    {
+        /// <summary>
+        /// Gets the specified cacheKey, dataRetriever and expiration async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        /// <param name="dataRetriever">Data retriever.</param>
+        /// <param name="expiration">Expiration.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey, Func<Task<T>> dataRetriever,
+            TimeSpan expiration)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
+
+            var result = _cache.Get<T>(cacheKey);
+            if (result.HasValue)
+            {
+                if (_options.EnableLogging)
+                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
+
+                CacheStats.OnHit();
+
+                return result;
+            }
+
+            CacheStats.OnMiss();
+
+            if (_options.EnableLogging)
+                _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
+
+            if (!_cache.Add($"{cacheKey}_Lock", 1, TimeSpan.FromMilliseconds(_options.LockMs)))
+            {
+                //wait for some ms
+                await Task.Delay(_options.SleepMs);
+                return await GetAsync(cacheKey, dataRetriever, expiration);
+            }
+
+            var res = await dataRetriever();
+
+            if (res != null)
+            {
+                await SetAsync(cacheKey, res, expiration);
+                //remove mutex key
+                _cache.Remove($"{cacheKey}_Lock");
+
+                return new CacheValue<T>(res, true);
+            }
+            else
+            {
+                //remove mutex key
+                _cache.Remove($"{cacheKey}_Lock");
+                return CacheValue<T>.NoValue;
+            }
+        }
+
+        /// <summary>
+        /// Gets the specified cacheKey async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+
+            var result = await Task.FromResult(_cache.Get<T>(cacheKey));
+
+            if (result.HasValue)
+            {
+                if (_options.EnableLogging)
+                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
+
+                CacheStats.OnHit();
+
+                return result;
+            }
+            else
+            {
+                if (_options.EnableLogging)
+                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
+
+                CacheStats.OnMiss();
+
+                return CacheValue<T>.NoValue;
+            }
+        }
+
+        /// <summary>
+        /// Gets the specified cacheKey async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        /// <param name="type">Object Type.</param>
+        public override async Task<object> BaseGetAsync(string cacheKey, Type type)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+
+            var result = await Task.FromResult(_cache.Get(cacheKey));
+
+            if (result != null)
+            {
+                if (_options.EnableLogging)
+                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
+
+                CacheStats.OnHit();
+
+                return result;
+            }
+            else
+            {
+                if (_options.EnableLogging)
+                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
+
+                CacheStats.OnMiss();
+
+                return null;
+            }
+        }
+
+        /// <summary>
+        /// Removes the specified cacheKey async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        public override async Task BaseRemoveAsync(string cacheKey)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+
+            await Task.Run(() => { _cache.Remove(cacheKey); });
+        }      
+
+        /// <summary>
+        /// Sets the specified cacheKey, cacheValue and expiration async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        /// <param name="cacheValue">Cache value.</param>
+        /// <param name="expiration">Expiration.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override async Task BaseSetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
+            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
+
+            if (MaxRdSecond > 0)
+            {
+                var addSec = new Random().Next(1, MaxRdSecond);
+                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
+            }
+
+            await Task.Run(() =>
+            {
+                //var valExpiration = expiration.Seconds <= 1 ? expiration : TimeSpan.FromSeconds(expiration.Seconds / 2);
+                //var val = new CacheValue<T>(cacheValue, true, valExpiration);
+                _cache.Set(cacheKey, cacheValue, expiration);
+            });
+        }
+
+        /// <summary>
+        /// Existses the specified cacheKey async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        public override async Task<bool> BaseExistsAsync(string cacheKey)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+
+            return await Task.FromResult(_cache.Exists(cacheKey));
+        }              
+
+        /// <summary>
+        /// Removes cached item by cachekey's prefix async.
+        /// </summary>
+        /// <returns>The by prefix async.</returns>
+        /// <param name="prefix">Prefix.</param>
+        public override async Task BaseRemoveByPrefixAsync(string prefix)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
+
+            var count = await Task.Run(() => _cache.RemoveByPrefix(prefix));
+
+            if (_options.EnableLogging)
+                _logger?.LogInformation($"RemoveByPrefixAsync : prefix = {prefix} , count = {count}");
+        }
+      
+        /// <summary>
+        /// Sets all async.
+        /// </summary>
+        /// <returns>The all async.</returns>
+        /// <param name="values">Values.</param>
+        /// <param name="expiration">Expiration.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override async Task BaseSetAllAsync<T>(IDictionary<string, T> values, TimeSpan expiration)
+        {
+            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
+            ArgumentCheck.NotNullAndCountGTZero(values, nameof(values));
+
+            await Task.Run(() => _cache.SetAll(values, expiration));
+        }
+
+        /// <summary>
+        /// Gets all async.
+        /// </summary>
+        /// <returns>The all async.</returns>
+        /// <param name="cacheKeys">Cache keys.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetAllAsync<T>(IEnumerable<string> cacheKeys)
+        {
+            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
+
+            if (_options.EnableLogging)
+                _logger?.LogInformation($"GetAllAsync : cacheKeys = {string.Join(",", cacheKeys)}");
+
+            return await Task.FromResult(_cache.GetAll<T>(cacheKeys));
+        }
+      
+        /// <summary>
+        /// Gets the by prefix async.
+        /// </summary>
+        /// <returns>The by prefix async.</returns>
+        /// <param name="prefix">Prefix.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetByPrefixAsync<T>(string prefix)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
+            var map = new Dictionary<string, CacheValue<T>>();
+
+            if (_options.EnableLogging)
+                _logger?.LogInformation($"GetByPrefixAsync : prefix = {prefix}");
+
+            return await Task.FromResult(_cache.GetByPrefix<T>(prefix));
+        }
+
+        /// <summary>
+        /// Removes all async.
+        /// </summary>
+        /// <returns>The all async.</returns>
+        /// <param name="cacheKeys">Cache keys.</param>
+        public override async Task BaseRemoveAllAsync(IEnumerable<string> cacheKeys)
+        {
+            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
+
+            if (_options.EnableLogging)
+                _logger?.LogInformation($"RemoveAllAsync : cacheKeys = {string.Join(",", cacheKeys)}");
+
+            await Task.Run(() => _cache.RemoveAll(cacheKeys));
+        }
+
+        /// <summary>
+        /// Flush All Cached Item async.
+        /// </summary>
+        /// <returns>The async.</returns>
+        public override async Task BaseFlushAsync()
+        {
+            if (_options.EnableLogging)
+                _logger?.LogInformation("FlushAsync");
+
+            _cache.Clear();
+            await Task.CompletedTask;
+        }
+
+        /// <summary>
+        /// Tries the set async.
+        /// </summary>
+        /// <returns>The set async.</returns>
+        /// <param name="cacheKey">Cache key.</param>
+        /// <param name="cacheValue">Cache value.</param>
+        /// <param name="expiration">Expiration.</param>
+        /// <typeparam name="T">The 1st type parameter.</typeparam>
+        public override Task<bool> BaseTrySetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
+            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
+
+            //var val = new CacheValue<T>(cacheValue, true, expiration);
+            return Task.FromResult(_cache.Add(cacheKey, cacheValue, expiration));
+        }
+
+        /// <summary>
+        /// Get the expiration of cache key
+        /// </summary>
+        /// <param name="cacheKey">cache key</param>
+        /// <returns>expiration</returns>
+        public override Task<TimeSpan> BaseGetExpirationAsync(string cacheKey)
+        {
+            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
+
+            return Task.FromResult(_cache.GetExpiration(cacheKey));
+        }     
+    }
+}
