+++ src/EasyCaching.Memcached/DefaultMemcachedCachingProvider.cs
@@ -1,19 +1,17 @@
 ﻿namespace EasyCaching.Memcached
 {
-    using EasyCaching.Core;
-    using Microsoft.Extensions.Logging;
-    using Microsoft.Extensions.Options;
     using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Security.Cryptography;
     using System.Text;
-    using System.Threading.Tasks;
+    using EasyCaching.Core;
+    using Microsoft.Extensions.Logging;
 
     /// <summary>
     /// Default memcached caching provider.
     /// </summary>
-    public class DefaultMemcachedCachingProvider : EasyCachingAbstractProvider
+    public partial class DefaultMemcachedCachingProvider : EasyCachingAbstractProvider
     {
         /// <summary>
         /// The memcached client.
@@ -130,58 +128,7 @@
                 return CacheValue<T>.NoValue;
             }
         }
-
-        /// <summary>
-        /// Gets the specified cacheKey, dataRetriever and expiration async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="dataRetriever">Data retriever.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey, Func<Task<T>> dataRetriever, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            var result = await _memcachedClient.GetAsync<T>(this.HandleCacheKey(cacheKey));
-            if (result.Success)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                return new CacheValue<T>(result.Value, true);
-            }
-
-            CacheStats.OnMiss();
-
-            if (_options.EnableLogging)
-                _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-            var flag = await _memcachedClient.StoreAsync(Enyim.Caching.Memcached.StoreMode.Add, this.HandleCacheKey($"{cacheKey}_Lock"), 1, TimeSpan.FromMilliseconds(_options.LockMs));
-
-            if (!flag)
-            {
-                await Task.Delay(_options.SleepMs);
-                return await GetAsync(cacheKey, dataRetriever, expiration);
-            }
-
-            var item = await dataRetriever();
-            if (item != null)
-            {
-                await this.SetAsync(cacheKey, item, expiration);
-                await _memcachedClient.RemoveAsync(this.HandleCacheKey($"{cacheKey}_Lock"));
-                return new CacheValue<T>(item, true);
-            }
-            else
-            {
-                await _memcachedClient.RemoveAsync(this.HandleCacheKey($"{cacheKey}_Lock"));
-                return CacheValue<T>.NoValue;
-            }
-        }
-
+    
         /// <summary>
         /// Get the specified cacheKey.
         /// </summary>
@@ -213,68 +160,6 @@
         }
 
         /// <summary>
-        /// Gets the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var result = await _memcachedClient.GetAsync<T>(this.HandleCacheKey(cacheKey));
-            if (result.Success)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                return new CacheValue<T>(result.Value, true);
-            }
-            else
-            {
-                CacheStats.OnMiss();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-                return CacheValue<T>.NoValue;
-            }
-        }
-
-        /// <summary>
-        /// Gets the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="type">Object Type.</param>
-        public override async Task<object> BaseGetAsync(string cacheKey, Type type)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var result = await Task.FromResult(_memcachedClient.Get(this.HandleCacheKey(cacheKey)));
-            if (result != null)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                return result;
-            }
-            else
-            {
-                CacheStats.OnMiss();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-                return null;
-            }
-        }
-
-        /// <summary>
         /// Remove the specified cacheKey.
         /// </summary>
         /// <returns>The remove.</returns>
@@ -284,19 +169,7 @@
             ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
 
             _memcachedClient.Remove(this.HandleCacheKey(cacheKey));
-        }
-
-        /// <summary>
-        /// Removes the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        public override async Task BaseRemoveAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            await _memcachedClient.RemoveAsync(this.HandleCacheKey(cacheKey));
-        }
+        }    
 
         /// <summary>
         /// Set the specified cacheKey, cacheValue and expiration.
@@ -320,30 +193,7 @@
 
             _memcachedClient.Store(Enyim.Caching.Memcached.StoreMode.Set, this.HandleCacheKey(cacheKey), cacheValue, expiration);
         }
-
-        /// <summary>
-        /// Sets the specified cacheKey, cacheValue and expiration async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="cacheValue">Cache value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task BaseSetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            await _memcachedClient.StoreAsync(Enyim.Caching.Memcached.StoreMode.Set, this.HandleCacheKey(cacheKey), cacheValue, expiration);
-        }
-
+      
         /// <summary>
         /// Exists the specified cacheKey.
         /// </summary>
@@ -356,18 +206,6 @@
             return _memcachedClient.TryGet(this.HandleCacheKey(cacheKey), out object obj);
         }
 
-        /// <summary>
-        /// Existses the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        public override async Task<bool> BaseExistsAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            return await Task.FromResult(_memcachedClient.TryGet(this.HandleCacheKey(cacheKey), out object obj));
-        }
-              
         /// <summary>
         /// Removes cached item by cachekey's prefix.
         /// </summary>
@@ -394,35 +232,7 @@
             }
             _memcachedClient.Store(Enyim.Caching.Memcached.StoreMode.Set, this.HandleCacheKey(prefix), newValue, new TimeSpan(0, 0, 0));
         }
-
-        /// <summary>
-        /// Removes cached item by cachekey's prefix async.
-        /// </summary>
-        /// <remarks>
-        /// Before using the method , you should follow this link 
-        /// https://github.com/memcached/memcached/wiki/ProgrammingTricks#namespacing
-        /// and confirm that you use the namespacing when you set and get the cache.
-        /// </remarks>
-        /// <param name="prefix">Prefix of CacheKey.</param>
-        /// <returns></returns>
-        public override async Task BaseRemoveByPrefixAsync(string prefix)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
-
-            var oldPrefixKey = _memcachedClient.Get(prefix)?.ToString();
-
-            var newValue = DateTime.UtcNow.Ticks.ToString();
-
-            if (_options.EnableLogging)
-                _logger?.LogInformation($"RemoveByPrefixAsync : prefix = {prefix}");
-
-            if (oldPrefixKey.Equals(newValue))
-            {
-                newValue = string.Concat(newValue, new Random().Next(9).ToString());
-            }
-            await _memcachedClient.StoreAsync(Enyim.Caching.Memcached.StoreMode.Set, this.HandleCacheKey(prefix), newValue, new TimeSpan(0, 0, 0));
-        }
-
+      
         /// <summary>
         /// Handle the cache key of memcached limititaion
         /// </summary>
@@ -460,27 +270,7 @@
                 Set(item.Key, item.Value, expiration);
             }
         }
-
-        /// <summary>
-        /// Sets all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="values">Values.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task BaseSetAllAsync<T>(IDictionary<string, T> values, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-            ArgumentCheck.NotNullAndCountGTZero(values, nameof(values));
-
-            var tasks = new List<Task>();
-            foreach (var item in values)
-            {
-                tasks.Add(SetAsync(item.Key, item.Value, expiration));
-            }
-            await Task.WhenAll(tasks);
-        }
-
+       
         /// <summary>
         /// Gets all.
         /// </summary>
@@ -504,31 +294,7 @@
 
             return result;
         }
-
-        /// <summary>
-        /// Gets all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="cacheKeys">Cache keys.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetAllAsync<T>(IEnumerable<string> cacheKeys)
-        {
-            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
-
-            var values = await _memcachedClient.GetAsync<T>(cacheKeys);
-            var result = new Dictionary<string, CacheValue<T>>();
-
-            foreach (var item in values)
-            {
-                if (item.Value != null)
-                    result.Add(item.Key, new CacheValue<T>(item.Value, true));
-                else
-                    result.Add(item.Key, CacheValue<T>.NoValue);
-            }
-
-            return result;
-        }
-
+     
         /// <summary>
         /// Gets the by prefix.
         /// </summary>
@@ -536,17 +302,6 @@
         /// <param name="prefix">Prefix.</param>
         /// <typeparam name="T">The 1st type parameter.</typeparam>
         public override IDictionary<string, CacheValue<T>> BaseGetByPrefix<T>(string prefix)
-        {
-            throw new NotImplementedException();
-        }
-
-        /// <summary>
-        /// Gets the by prefix async.
-        /// </summary>
-        /// <returns>The by prefix async.</returns>
-        /// <param name="prefix">Prefix.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override Task<IDictionary<string, CacheValue<T>>> BaseGetByPrefixAsync<T>(string prefix)
         {
             throw new NotImplementedException();
         }
@@ -562,23 +317,7 @@
             foreach (var item in cacheKeys.Distinct())
                 Remove(item);
         }
-
-        /// <summary>
-        /// Removes all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="cacheKeys">Cache keys.</param>
-        public override async Task BaseRemoveAllAsync(IEnumerable<string> cacheKeys)
-        {
-            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
-
-            var tasks = new List<Task>();
-            foreach (var item in cacheKeys.Distinct())
-                tasks.Add(RemoveAsync(item));
-
-            await Task.WhenAll(tasks);
-        }
-
+    
         /// <summary>
         /// Gets the count.
         /// </summary>
@@ -610,18 +349,6 @@
         }
 
         /// <summary>
-        /// Flush All Cached Item async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        public override async Task BaseFlushAsync()
-        {
-            if (_options.EnableLogging)
-                _logger?.LogInformation("Memcached -- FlushAsync");
-
-            await _memcachedClient.FlushAllAsync();
-        }
-
-        /// <summary>
         /// Tries the set.
         /// </summary>
         /// <returns><c>true</c>, if set was tryed, <c>false</c> otherwise.</returns>
@@ -644,39 +371,15 @@
             return _memcachedClient.Store(Enyim.Caching.Memcached.StoreMode.Add, this.HandleCacheKey(cacheKey), cacheValue, expiration);
         }
 
-        /// <summary>
-        /// Tries the set async.
-        /// </summary>
-        /// <returns>The set async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="cacheValue">Cache value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override Task<bool> BaseTrySetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            return _memcachedClient.StoreAsync(Enyim.Caching.Memcached.StoreMode.Add, this.HandleCacheKey(cacheKey), cacheValue, expiration);
-        }
-
         public override TimeSpan BaseGetExpiration(string cacheKey)
         {
             throw new NotImplementedException();
         }
 
-        public override Task<TimeSpan> BaseGetExpirationAsync(string cacheKey)
-        {
-            throw new NotImplementedException();
-        }
-
+        /// <summary>
+        /// Get te information of this provider.
+        /// </summary>
+        /// <returns></returns>   
         public override ProviderInfo BaseGetProviderInfo()
         {
             return _info;
