+++ src/EasyCaching.Redis/DefaultRedisCachingProvider.cs
@@ -3,11 +3,9 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using System.Threading.Tasks;
     using EasyCaching.Core;
     using EasyCaching.Core.Serialization;
     using Microsoft.Extensions.Logging;
-    using Microsoft.Extensions.Options;
     using StackExchange.Redis;
 
     /// <summary>
@@ -161,93 +159,6 @@
         }
 
         /// <summary>
-        /// Gets the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="type">Object Type.</param>
-        public override async Task<object> BaseGetAsync(string cacheKey, Type type)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var result = await _cache.StringGetAsync(cacheKey);
-            if (!result.IsNull)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                var value = _serializer.Deserialize(result, type);
-                return value;
-            }
-            else
-            {
-                CacheStats.OnMiss();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Gets the specified cacheKey, dataRetriever and expiration async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="dataRetriever">Data retriever.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey, Func<Task<T>> dataRetriever, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            var result = await _cache.StringGetAsync(cacheKey);
-            if (!result.IsNull)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                var value = _serializer.Deserialize<T>(result);
-                return new CacheValue<T>(value, true);
-            }
-
-            CacheStats.OnMiss();
-
-            if (_options.EnableLogging)
-                _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-            var flag = await _cache.StringSetAsync($"{cacheKey}_Lock", 1, TimeSpan.FromMilliseconds(_options.LockMs), When.NotExists);
-
-            if (!flag)
-            {
-                await Task.Delay(_options.SleepMs);
-                return await GetAsync(cacheKey, dataRetriever, expiration);
-            }
-
-            var item = await dataRetriever();
-            if (item != null)
-            {
-                await SetAsync(cacheKey, item, expiration);
-
-                //remove mutex key
-                await _cache.KeyDeleteAsync($"{cacheKey}_Lock");
-                return new CacheValue<T>(item, true);
-            }
-            else
-            {
-                //remove mutex key
-                await _cache.KeyDeleteAsync($"{cacheKey}_Lock");
-                return CacheValue<T>.NoValue;
-            }
-        }
-
-        /// <summary>
         /// Get the specified cacheKey.
         /// </summary>
         /// <returns>The get.</returns>
@@ -258,38 +169,6 @@
             ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
 
             var result = _cache.StringGet(cacheKey);
-            if (!result.IsNull)
-            {
-                CacheStats.OnHit();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Hit : cachekey = {cacheKey}");
-
-                var value = _serializer.Deserialize<T>(result);
-                return new CacheValue<T>(value, true);
-            }
-            else
-            {
-                CacheStats.OnMiss();
-
-                if (_options.EnableLogging)
-                    _logger?.LogInformation($"Cache Missed : cachekey = {cacheKey}");
-
-                return CacheValue<T>.NoValue;
-            }
-        }
-
-        /// <summary>
-        /// Gets the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<CacheValue<T>> BaseGetAsync<T>(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var result = await _cache.StringGetAsync(cacheKey);
             if (!result.IsNull)
             {
                 CacheStats.OnHit();
@@ -324,18 +203,6 @@
         }
 
         /// <summary>
-        /// Removes the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        public override async Task BaseRemoveAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            await _cache.KeyDeleteAsync(cacheKey);
-        }
-
-        /// <summary>
         /// Set the specified cacheKey, cacheValue and expiration.
         /// </summary>
         /// <returns>The set.</returns>
@@ -362,32 +229,6 @@
         }
 
         /// <summary>
-        /// Sets the specified cacheKey, cacheValue and expiration async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="cacheValue">Cache value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task BaseSetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            await _cache.StringSetAsync(
-                    cacheKey,
-                    _serializer.Serialize(cacheValue),
-                    expiration);
-        }
-
-        /// <summary>
         /// Exists the specified cacheKey.
         /// </summary>
         /// <returns>The exists.</returns>
@@ -397,18 +238,6 @@
             ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
 
             return _cache.KeyExists(cacheKey);
-        }
-
-        /// <summary>
-        /// Existses the specified cacheKey async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        public override async Task<bool> BaseExistsAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            return await _cache.KeyExistsAsync(cacheKey);
         }
 
         /// <summary>
@@ -427,24 +256,6 @@
             var redisKeys = this.SearchRedisKeys(prefix);
 
             _cache.KeyDelete(redisKeys);
-        }
-
-        /// <summary>
-        /// Removes cached item by cachekey's prefix async.
-        /// </summary>
-        /// <param name="prefix">Prefix of CacheKey.</param>
-        public override async Task BaseRemoveByPrefixAsync(string prefix)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
-
-            prefix = this.HandlePrefix(prefix);
-
-            if (_options.EnableLogging)
-                _logger?.LogInformation($"RemoveByPrefixAsync : prefix = {prefix}");
-
-            var redisKeys = this.SearchRedisKeys(prefix);
-
-            await _cache.KeyDeleteAsync(redisKeys);
         }
 
         /// <summary>
@@ -527,26 +338,6 @@
         }
 
         /// <summary>
-        /// Sets all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="values">Values.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task BaseSetAllAsync<T>(IDictionary<string, T> values, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-            ArgumentCheck.NotNullAndCountGTZero(values, nameof(values));
-
-            var tasks = new List<Task>();
-
-            foreach (var item in values)
-                tasks.Add(SetAsync(item.Key, item.Value, expiration));
-
-            await Task.WhenAll(tasks);
-        }
-
-        /// <summary>
         /// Gets all.
         /// </summary>
         /// <returns>The all.</returns>
@@ -558,32 +349,6 @@
 
             var keyArray = cacheKeys.ToArray();
             var values = _cache.StringGet(keyArray.Select(k => (RedisKey)k).ToArray());
-
-            var result = new Dictionary<string, CacheValue<T>>();
-            for (int i = 0; i < keyArray.Length; i++)
-            {
-                var cachedValue = values[i];
-                if (!cachedValue.IsNull)
-                    result.Add(keyArray[i], new CacheValue<T>(_serializer.Deserialize<T>(cachedValue), true));
-                else
-                    result.Add(keyArray[i], CacheValue<T>.NoValue);
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Gets all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="cacheKeys">Cache keys.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetAllAsync<T>(IEnumerable<string> cacheKeys)
-        {
-            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
-
-            var keyArray = cacheKeys.ToArray();
-            var values = await _cache.StringGetAsync(keyArray.Select(k => (RedisKey)k).ToArray());
 
             var result = new Dictionary<string, CacheValue<T>>();
             for (int i = 0; i < keyArray.Length; i++)
@@ -628,35 +393,6 @@
         }
 
         /// <summary>
-        /// Gets the by prefix async.
-        /// </summary>
-        /// <returns>The by prefix async.</returns>
-        /// <param name="prefix">Prefix.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override async Task<IDictionary<string, CacheValue<T>>> BaseGetByPrefixAsync<T>(string prefix)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(prefix, nameof(prefix));
-
-            prefix = this.HandlePrefix(prefix);
-
-            var redisKeys = this.SearchRedisKeys(prefix);
-
-            var values = (await _cache.StringGetAsync(redisKeys)).ToArray();
-
-            var result = new Dictionary<string, CacheValue<T>>();
-            for (int i = 0; i < redisKeys.Length; i++)
-            {
-                var cachedValue = values[i];
-                if (!cachedValue.IsNull)
-                    result.Add(redisKeys[i], new CacheValue<T>(_serializer.Deserialize<T>(cachedValue), true));
-                else
-                    result.Add(redisKeys[i], CacheValue<T>.NoValue);
-            }
-
-            return result;
-        }
-
-        /// <summary>
         /// Removes all.
         /// </summary>
         /// <param name="cacheKeys">Cache keys.</param>
@@ -667,20 +403,6 @@
             var redisKeys = cacheKeys.Where(k => !string.IsNullOrEmpty(k)).Select(k => (RedisKey)k).ToArray();
             if (redisKeys.Length > 0)
                 _cache.KeyDelete(redisKeys);
-        }
-
-        /// <summary>
-        /// Removes all async.
-        /// </summary>
-        /// <returns>The all async.</returns>
-        /// <param name="cacheKeys">Cache keys.</param>
-        public override async Task BaseRemoveAllAsync(IEnumerable<string> cacheKeys)
-        {
-            ArgumentCheck.NotNullAndCountGTZero(cacheKeys, nameof(cacheKeys));
-
-            var redisKeys = cacheKeys.Where(k => !string.IsNullOrEmpty(k)).Select(k => (RedisKey)k).ToArray();
-            if (redisKeys.Length > 0)
-                await _cache.KeyDeleteAsync(redisKeys);
         }
 
         /// <summary>
@@ -718,25 +440,6 @@
         }
 
         /// <summary>
-        /// Flush All Cached Item async.
-        /// </summary>
-        /// <returns>The async.</returns>
-        public override async Task BaseFlushAsync()
-        {
-            if (_options.EnableLogging)
-                _logger?.LogInformation("Redis -- FlushAsync");
-
-            var tasks = new List<Task>();
-
-            foreach (var server in _servers)
-            {
-                tasks.Add(server.FlushDatabaseAsync(_cache.Database));
-            }
-
-            await Task.WhenAll(tasks);
-        }
-
-        /// <summary>
         /// Tries the set.
         /// </summary>
         /// <returns><c>true</c>, if set was tryed, <c>false</c> otherwise.</returns>
@@ -765,33 +468,10 @@
         }
 
         /// <summary>
-        /// Tries the set async.
+        /// Get the expiration of cache key
         /// </summary>
-        /// <returns>The set async.</returns>
-        /// <param name="cacheKey">Cache key.</param>
-        /// <param name="cacheValue">Cache value.</param>
-        /// <param name="expiration">Expiration.</param>
-        /// <typeparam name="T">The 1st type parameter.</typeparam>
-        public override Task<bool> BaseTrySetAsync<T>(string cacheKey, T cacheValue, TimeSpan expiration)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-            ArgumentCheck.NotNull(cacheValue, nameof(cacheValue));
-            ArgumentCheck.NotNegativeOrZero(expiration, nameof(expiration));
-
-            if (MaxRdSecond > 0)
-            {
-                var addSec = new Random().Next(1, MaxRdSecond);
-                expiration = expiration.Add(TimeSpan.FromSeconds(addSec));
-            }
-
-            return _cache.StringSetAsync(
-                cacheKey,
-                _serializer.Serialize(cacheValue),
-                expiration,
-                When.NotExists
-                );
-        }
-
+        /// <param name="cacheKey">cache key</param>
+        /// <returns>expiration</returns>
         public override TimeSpan BaseGetExpiration(string cacheKey)
         {
             ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
@@ -800,14 +480,10 @@
             return timeSpan.HasValue ? timeSpan.Value : TimeSpan.Zero;
         }
 
-        public override async Task<TimeSpan> BaseGetExpirationAsync(string cacheKey)
-        {
-            ArgumentCheck.NotNullOrWhiteSpace(cacheKey, nameof(cacheKey));
-
-            var timeSpan = await _cache.KeyTimeToLiveAsync(cacheKey);
-            return timeSpan.HasValue ? timeSpan.Value : TimeSpan.Zero;
-        }
-
+        /// <summary>
+        /// Get te information of this provider.
+        /// </summary>
+        /// <returns></returns>
         public override ProviderInfo BaseGetProviderInfo()
         {
             return _info;
